% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_conf_p_values.R
\name{agg_conf_p_values}
\alias{agg_conf_p_values}
\title{Outputs the sce returned by conformeR with gene-level aggregated p-values within each conformal groups. Averaging is performed.}
\usage{
agg_conf_p_values(sce)
}
\arguments{
\item{sce}{output of conformeR with the assay "adj_pvalues".}
}
\value{
the sce returned by conformeR with gene-level aggregated p-values within each conformal groups.
}
\description{
Outputs the sce returned by conformeR with gene-level aggregated p-values within each conformal groups. Averaging is performed.
}
