% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_pred.R
\name{cf_pred}
\alias{cf_pred}
\title{Imputation of counterfactual scenario and computation of conformal p-values for each gene x patient id.}
\usage{
cf_pred(
  sce,
  obs_condition,
  spacing = 0.1,
  prop_train = 0.5,
  prop_proper = 0.75,
  cores = 32
)
}
\arguments{
\item{sce}{an original sc-RNA-seq dataset of class SingleCellExperiment with obs_condition.}

\item{obs_condition}{name of the column in colData(sce) for the cell's observed condition specified as a string.}

\item{spacing}{level of precision for alpha's grid.}

\item{prop_train}{size in proportion of original data set for training set.}

\item{prop_proper}{size in proportion of training data set for proper set.}

\item{cores}{number of workers for BiocParallel job.}
}
\value{
a tibble of size n_cells x n_genes containing conformal p-values.
}
\description{
Imputation of counterfactual scenario and computation of conformal p-values for each gene x patient id.
}
