% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_pval.R
\name{conf_pval}
\alias{conf_pval}
\title{Outputs a p-value encoding the difference between the observed and predicted expression levels for a new observed count.
All data given as input has to refer to a same gene and belonging to cells within a same conformal group.}
\usage{
conf_pval(obs_train, pred_train, obs_cal, pred_cal, obs_test, alpha = 0.1)
}
\arguments{
\item{obs_train}{a vector of observed logcounts, on which the quantile regression forest is trained.}

\item{pred_train}{a vector of predicted lemur logcounts of the same length as obs_train.}

\item{obs_cal}{a vector of observed logcounts, to be passed as argument of the score function.}

\item{pred_cal}{a vector of predicted lemur logcounts of the same length as obs_cal.}

\item{obs_test}{a new observed logcount.}

\item{alpha}{level of confidence between 0 and 1.}
}
\value{
a single p-value encoding the difference between the observed and predicted expression levels for a new observed count.
}
\description{
Outputs a p-value encoding the difference between the observed and predicted expression levels for a new observed count.
All data given as input has to refer to a same gene and belonging to cells within a same conformal group.
}
