% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformeR.R
\name{conformeR}
\alias{conformeR}
\title{LEMUR analysis for neighborhoods on predicted gene differential expression with conformal methods}
\usage{
conformeR(
  sce,
  design_lemur = ~1,
  contrast_column,
  contrast_levels = NULL,
  n_embedding = 60,
  use_assay = "logcounts",
  test_fraction_lemur = 0.2,
  cp_train_frac = 0.5,
  cp_cal_frac = 0.25,
  cp_alpha = 0.05,
  genes_of_interest,
  what = c("conf_selection", "conf_clustering"),
  n_cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object containing the single-cell
expression data.}

\item{design_lemur}{A formula specifying the design used by LEMUR.}

\item{contrast_column}{A column name specifying the comparison used for
differential-expression testing in LEMUR.}

\item{n_embedding}{Number of dimensions in the LEMUR embedding.}

\item{use_assay}{Character string specifying the assay to use.}

\item{test_fraction_lemur}{Fraction of cells used as the LEMUR test set.}

\item{cp_train_frac}{Fraction of cells used for conformal training.}

\item{cp_cal_frac}{Fraction of cells used for conformal calibration.}

\item{cp_alpha}{Miscoverage level for the conformal procedures.}

\item{genes_of_interest}{Character vector containing the names of genes
to include in the conformal analysis.They need to correspond to the row names of the sce.}

\item{what}{Character vector specifying the conformal procedure(s) to run.}

\item{n_cores}{Number of cores used for parallel processing.}

\item{verbose}{Logical indicating whether to display progress information.}

\item{constrast_levels}{when \code{contrast_column} has more than two levels, specify which two are tested again each other.}
}
\value{
A list containing:
\describe{
\item{\code{fit_lemur}}{The LEMUR test-set object restricted to the
genes of interest.}
\item{\code{nei_lemur}}{The LEMUR neighborhoods for the genes of
interest in the test set.}
\item{\code{conf_results}}{The results of the conformal procedures
for the genes of interest.}
}
}
\description{
Runs a LEMUR analysis, identifies differential-expression neighborhoods,
and applies conformal procedures to a set of genes of interest.
The data are split into training, calibration, and test sets based on the
covariates specified in the LEMUR design formula.
}
\details{
The conformal procedures are applied to a subset of genes of
interest. Genes are processed in chunks to allow parallel computation.
}
