% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformeR_cf.R
\name{conformeR_cf}
\alias{conformeR_cf}
\title{Prediction interval builder for difference in conditions, based on counterfactual scenario prediction.}
\usage{
conformeR_cf(
  sce,
  obs_condition,
  replicate_id,
  cell_type,
  spacing = 0.01,
  size_train = 0.5,
  size_cal = 0.25,
  cores = 32,
  gene_chunk = 200
)
}
\arguments{
\item{sce}{SingleCellExperiment with replicate_id, obs_condition, cell_type in colData}

\item{obs_condition}{column name for observed condition (string)}

\item{replicate_id}{column name for biological replicate (string)}

\item{cell_type}{column name for cell types (string)}

\item{spacing}{level of precision for alpha's grid.}

\item{size_train}{proportion for training set.}

\item{size_cal}{proportion of training set for calibration set.}

\item{cores}{number of workers for BiocParallel job.}
}
\value{
a data frame of prediction intervals for difference in observed vs. counterfactual logcount for each observed gene x cell x confidence level
}
\description{
Prediction interval builder for difference in conditions, based on counterfactual scenario prediction.
}
