% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformeR_cfcausal.R
\name{conformeR_cfcausal}
\alias{conformeR_cfcausal}
\title{Imputation of counterfactual scenario and computation of conformal p-values for each gene x patient id.}
\usage{
conformeR_cfcausal(
  sce,
  obs_condition,
  replicate_id,
  cell_type,
  spacing = 0.01,
  size_train = 0.5,
  cores = 32
)
}
\arguments{
\item{sce}{SingleCellExperiment with replicate_id, obs_condition, cell_type in colData}

\item{obs_condition}{column name for observed condition (string)}

\item{replicate_id}{column name for biological replicate (string)}

\item{cell_type}{column name for cell types (string)}

\item{spacing}{level of precision for alpha's grid.}

\item{size_train}{proportion for training set.}

\item{cores}{number of workers for BiocParallel job.}

\item{cutoff}{confidence level for FDR}

\item{size_cal}{proportion for calibration set.}
}
\value{
a data frame of dimension number of \code{conf_group} x number of genes containing FDR(level=\code{cutoff}) for each gene in a given  \code{conf_group}
}
\description{
Imputation of counterfactual scenario and computation of conformal p-values for each gene x patient id.
}
