% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{data_processing}
\alias{data_processing}
\title{Splitting of the original dataset using \code{initial_split} and balancing on covariates passed to \code{lemur}.}
\usage{
data_processing(sce, strat_by, size_train = 0.3, size_cal = 0.25)
}
\arguments{
\item{sce}{SingleCellExperiment with replicate_id, obs_condition, cell_type in colData}

\item{strat_by}{vector of column names in colData(sce) on which we stratify.}

\item{size_train}{proportion for training set}

\item{size_cal}{proportion for calibration set of training set.}
}
\value{
list with original sce, proper training, calibration, and test sets
}
\description{
Splitting of the original dataset using \code{initial_split} and balancing on covariates passed to \code{lemur}.
}
