% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr1.R
\name{fdr1}
\alias{fdr1}
\title{Computation of FDR at level \code{cutoff} for each \code{conf_group}}
\usage{
fdr1(p_val_df, cutoff = 0.05, cores = 32)
}
\arguments{
\item{p_val_df}{data frame of conformal p-values as output by function \code{conformeR_cfcausal}}

\item{cutoff}{confidence level for FDR}

\item{cores}{number of workers for BiocParallel job.}
}
\value{
a data frame of dimension number of \code{conf_group} x number of genes containing FDR(level=\code{cutoff}) for each gene in a given  \code{conf_group}
}
\description{
Computation of FDR at level \code{cutoff} for each \code{conf_group}
}
