% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr2.R
\name{fdr2}
\alias{fdr2}
\title{Computation of FDR at level \code{cutoff} for each \code{conf_group}}
\usage{
fdr2(p_val_df, cutoff = 0.05, cores = 32)
}
\arguments{
\item{p_val_df}{data frame of conformal p-values as output by function \code{conformeR_cfcausal}}

\item{cutoff}{confidence level for FDR}

\item{cores}{number of workers for BiocParallel job.}
}
\value{
a list of length \code{unique(p_val_df$gene)} each containing a list of length \code{unique(conf_group)}+1 of functions for computing tail-area FDR, and combined FDR across \code{unique(conf_group)} at level \code{cutoff}
}
\description{
Computation of FDR at level \code{cutoff} for each \code{conf_group}
}
