% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur_imputation.R
\name{lemur_imputation}
\alias{lemur_imputation}
\title{Lemur imputation of training and calibration set using lemur fit trained on training set.}
\usage{
lemur_imputation(data_processed, replicate_id, obs_condition, cell_type)
}
\arguments{
\item{data_processed}{an output of the function data_processing. If provided, it replaces train_set and cal_set.}

\item{replicate_id}{name of the column in colData() for biological replicate id specified as a string.}

\item{obs_condition}{name of the column in colData() for the cell's observed condition specified as a string.}

\item{cell_type}{name of the column in colData(sce) for cell-types specified as a string.}
}
\value{
a list of length 2 containing the lemur-imputed training matrix and the lemur-imputed calibration matrix (no observed conditions but imputed values alone).
}
\description{
Lemur imputation of training and calibration set using lemur fit trained on training set.
}
