% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sce_to_wide_tibble}
\alias{sce_to_wide_tibble}
\title{Convert a SingleCellExperiment to a wide tibble}
\usage{
sce_to_wide_tibble(sce, obs_condition)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{obs_condition}{Column name (string) for the observed condition in \code{colData}.}
}
\value{
A tibble in wide format, with cells as rows and genes, \code{colData} as columns.
}
\description{
This helper converts a SingleCellExperiment object into a wide-format tibble.
Meta-data columns (\code{conf_group} and the observed condition) are meFged together with
the gene expression matrix (counts).
}
