% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_vignette}
\alias{sce_vignette}
\title{Example single-cell dataset}
\format{
A \code{SingleCellExperiment} object.
}
\source{
Zhao, W., Dovas, A., Spinazzi, E., et al. (2021).
Deconvolution of cell type-specific drug responses in human tumor
tissue with single-cell RNA-seq. \emph{Genome Medicine}, 13, 82.
GEO accession: GSE148842.
}
\usage{
sce_vignette
}
\description{
A small SingleCellExperiment object used to demonstrate
the conformeR workflow.
10k cells and 3k genes for 2 patients, 2 conditions, 3 cell-types.
}
\details{
This object is a subset of the dataset from Zhao et al. (2021),
processed for use as an example dataset in \pkg{conformeR}.
}
\keyword{datasets}
