% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score function underlying the conformal prediction step.}
\usage{
score(obs_train, pred_train, obs_cal, pred_cal, alpha)
}
\arguments{
\item{obs_train}{a vector of observed logcounts for the same gene, and on which the quantile regression forest is trained.}

\item{pred_train}{a vector of predicted lemur logcounts for the same gene and of the same length as obs_train.}

\item{obs_cal}{a vector of observed logcounts for the same gene, to be passed as argument of the score function.}

\item{pred_cal}{a vector of predicted lemur logcounts for the same gene and of the same length as obs_cal.}

\item{alpha}{level of confidence between 0 and 1.}
}
\value{
a vector of score values of the same length as obs_cal.
}
\description{
Score function underlying the conformal prediction step.
}
