% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_by_condition}
\alias{subset_by_condition}
\title{Subset a tibble by a condition}
\usage{
subset_by_condition(tibble, condition, value)
}
\arguments{
\item{tibble}{A tibble or data frame.}

\item{condition}{Name of the column used for filtering (string).}

\item{value}{Value to match in the condition column.}
}
\value{
A filtered tibble.
}
\description{
This function filters rows of a tibble where a given column \code{condition}matches a \code{value}.
}
