% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformeR.R
\name{conformeR}
\alias{conformeR}
\title{Outputs a matrix of (adjusted) p-values encoding the difference between the observed and predicted expression levels for each pair gene x cell in a test set given training set and calibration set, all three randomly built from an original sc-RNA-seq dataset of class SingleCellExperiment with replicate_id, obs_condition, cell_type annotated in colData.}
\usage{
conformeR(
  sce,
  replicate_id,
  obs_condition,
  cell_type,
  size_train = 0.25,
  size_cal = 0.25,
  alpha = 0.1,
  cores = 32,
  file_name
)
}
\arguments{
\item{sce}{an original sc-RNA-seq dataset of class SingleCellExperiment with replicate_id, obs_condition, cell_type annotated in colData.}

\item{replicate_id}{name of the column in colData(sce) for biological replicate id specified as a string.}

\item{obs_condition}{name of the column in colData(sce) for the cell's observed condition specified as a string.}

\item{cell_type}{name of the column in colData(sce) for cell-types specified as a string.}

\item{size_train}{size in proportion of original data set for training set.}

\item{size_cal}{size in proportion of original data set for calibration set.}

\item{alpha}{level of confidence between 0 and 1.}

\item{cores}{number of workers for BiocParallel job.}
}
\value{
a test_set with two new assays named "pvalues" resp. "adj_pvalues" containing a matrix of p_values for each pair gene x cell resp. BH-adjusted p_values for each pair gene x cell..
}
\description{
Outputs a matrix of (adjusted) p-values encoding the difference between the observed and predicted expression levels for each pair gene x cell in a test set given training set and calibration set, all three randomly built from an original sc-RNA-seq dataset of class SingleCellExperiment with replicate_id, obs_condition, cell_type annotated in colData.
}
