% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{data_processing}
\alias{data_processing}
\title{\enumerate{
\item creation of conformal groups \code{conf_group}.
\item splitting of the original dataset using \code{initial_split} and balancing on \code{conf_group}.
}}
\usage{
data_processing(
  sce,
  replicate_id,
  obs_condition,
  cell_type,
  size_train = 0.5,
  size_cal = 0.25
)
}
\arguments{
\item{sce}{SingleCellExperiment with replicate_id, obs_condition, cell_type in colData}

\item{replicate_id}{column name for biological replicate (string)}

\item{obs_condition}{column name for observed condition (string)}

\item{cell_type}{column name for cell types (string)}

\item{size_train}{proportion for training set}

\item{size_cal}{proportion for calibration set of training set.}
}
\value{
list with original sce, proper training, calibration, and test sets
}
\description{
\enumerate{
\item creation of conformal groups \code{conf_group}.
\item splitting of the original dataset using \code{initial_split} and balancing on \code{conf_group}.
}
}
