% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{data_processing}
\alias{data_processing}
\title{Two-step data processing of the original dataset ;
\enumerate{
\item creation of conformal groups.
\item splitting of the original dataset
}}
\usage{
data_processing(
  sce,
  replicate_id,
  obs_condition,
  cell_type,
  size_train = 0.25,
  size_cal = 0.25
)
}
\arguments{
\item{sce}{an original sc-RNA-seq dataset of class SingleCellExperiment with replicate_id, obs_condition, cell_type annotated in colData.}

\item{replicate_id}{name of the column in colData(sce) for biological replicate id specified as a string.}

\item{obs_condition}{name of the column in colData(sce) for the cell's observed condition specified as a string.}

\item{cell_type}{name of the column in colData(sce) for cell-types specified as a string.}

\item{size_train}{size in proportion of original data set for training set.}

\item{size_cal}{size in proportion of original data set for calibration set.}
}
\value{
a list of SingleCellExperiment objects with original dataset, training set, calibration set, test set, including conformal groups id as a new column in colData()
}
\description{
Two-step data processing of the original dataset ;
\enumerate{
\item creation of conformal groups.
\item splitting of the original dataset
}
}
