% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_utils.R
\name{eta_conformeR}
\alias{eta_conformeR}
\title{Compute normalized weights and conformal threshold eta(x)}
\usage{
eta_conformeR(scores, weights_cal, weights_test, alpha = alpha)
}
\arguments{
\item{scores}{Numeric vector of conformal scores from the calibration set (output of \code{compute_cqr_scores}).}

\item{weights_cal}{Numeric vector of calibration weights.}

\item{weights_test}{Numeric scalar, weight associated with the test point.}

\item{alpha}{Numeric scalar, miscoverage level in (0,1).}
}
\value{
A numeric scalar giving the conformal threshold \eqn{\eta(x)}.
}
\description{
Internal helper function used in weighted split-CQR to normalize calibration
and test weights and compute the conformal threshold \eqn{\eta(x)}.
}
