% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr}
\alias{fdr}
\title{Computation of FDR at level \code{cutoff} for each \code{conf_group}}
\usage{
fdr(interval_df, cutoff = 0.05, cores = 32)
}
\arguments{
\item{interval_df}{data frame of prediction intervals as output by
function \code{conformeR_cfcausal} or \code{conformeR_cf}}

\item{cutoff}{confidence level for aggregated FDR}

\item{cores}{number of workers for BiocParallel job.}
}
\value{
a list with:
\itemize{
\item \code{comb_fdr}: tibble with gene-level combined FDR at cutoff
\item \code{group_fdr}: tibble with gene × conf_group FDR values at cutoff
\item \code{pvaldf}: the cell × gene p-values
}
}
\description{
Computation of FDR at level \code{cutoff} for each \code{conf_group}
}
