% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr}
\alias{fdr}
\title{Computation of FDR at level \code{cutoff} for each \code{conf_group}}
\usage{
fdr(interval_df, cutoff = 0.05, cores = 32)
}
\arguments{
\item{cutoff}{confidence level for aggregated FDR}

\item{cores}{number of workers for BiocParallel job.}

\item{A}{data frame of prediction intervals as output by function \code{conformeR_cfcausal} or \code{conformeR_cf}}
}
\value{
a list 3 with one p-values data-frame at level cell x gene,
one of length \code{unique(interval_df$gene)} each containing a list of length \code{unique(conf_group)}
of functions for computing group-wise tail-area FDR, one table of gene combined FDR across \code{unique(conf_group)} at level \code{cutoff}.
}
\description{
Computation of FDR at level \code{cutoff} for each \code{conf_group}
}
